requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
alembic==1.13.1
email-validator==2.1.0.post1
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
requests==2.31.0
firebase-admin>=6.0.0
pytest==8.0.0
httpx==0.26.0
black==24.1.1
flake8==7.0.0
isort==5.13.2